+++ app/src/main/java/com/gunnarro/android/terex/utility/Utility.java
@@ -28,7 +28,7 @@ import java.util.stream.Stream;
 public class Utility {
 
     public static final Integer DEFAULT_DAILY_BREAK_IN_MINUTES = 30;
-    public static final Long DEFAULT_DAILY_WORKING_HOURS_IN_MINUTES = 8 * 60L - DEFAULT_DAILY_BREAK_IN_MINUTES;
+    public static final Integer DEFAULT_DAILY_WORKING_HOURS_IN_MINUTES = 8 * 60 - DEFAULT_DAILY_BREAK_IN_MINUTES;
     private static final Pattern POSITIVE_INTEGER_PATTERN = Pattern.compile("\\d+");
     public static final Integer DEFAULT_HOURLY_RATE = 1250;
     private static final SimpleDateFormat dateFormatter;
